Update gradle support
